Refactor fee structure and student listing services for improved pagination; update validators to ensure consistent limit handling

+++ backend/validators/financeValidator.js
@@ -20,6 +20,6 @@ export const listPaymentsValidator = [
   query("date").optional().isISO8601().withMessage("Date must be valid"),
   query("startDate").optional().isISO8601().withMessage("Start date must be valid"),
   query("endDate").optional().isISO8601().withMessage("End date must be valid"),
-  query("page").optional().isInt({ min: 1 }).withMessage("Page must be a positive integer"),
-  query("limit").optional().isInt({ min: 1, max: 100 }).withMessage("Limit must be between 1 and 100"),
+  query("page").optional().isInt({ min: 1 }).toInt().withMessage("Page must be a positive integer"),
+  query("limit").optional().isInt({ min: 1, max: 100 }).toInt().withMessage("Limit must be between 1 and 100"),
 ];
